// EPS-Client/src/app/api/model/appointmentResponse.ts
/**
 * EPS API v1.0
 * API für das EPS
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LocationResponse } from './locationResponse';
import { PersonResponse } from './personResponse';
import { ProjectResponse } from './projectResponse';

export interface AppointmentResponse { 
    appointmentId: string;
    personResponse: PersonResponse;
    locationResponse: LocationResponse;
    projectResponse: ProjectResponse;
    endDate: Date;
    startDate: Date;
}