// EPS-Client/src/app/api/model/personResponse.ts
/**
 * EPS API v1.0
 * API für das EPS
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PersonResponse { 
    personId: string;
    firstname: string;
    lastname: string;
    eMail: string;
    dateOfBirth: Date;
}